[js, css] Added pledge bars + animation
I was trying to do this with the percentage, but the bars appears much smaller than what the design called for.

This need to be refactored for cases where the pledge amounts are so large the pixel values would cause the bars to go off the screen.

I would not call this production ready