[img, css] Updated checkbox styles for mobile spacing
The label would wrap around the pseudo checkbox on small widths. This did not look great. I have added flexbox around the wrapping element and removed the right margin from the  pseudo checkbox and used the widths of the containers to create the space from the design comp.

I also noticed I had a background on the checkbox SVG element. This has been removed.