[html, js, css] Fixed checkbox bug
The checkbox was not keeping in sync with the real checkbox it was wrapping because the click event was taking the label into consideration. When the label was clicked it would toggle the pseudo element, but the real checkbox would check itself and cause a desync.

I fixed this by keeping track of the pseudo elements click event. Then the label would be free to trigger it’s own click event to both the checkbox and pseudo checkbox elements